added config file and option to run as readonly in config file

+++ ClassLibrary/FileHandler.cs
@@ -10,13 +10,17 @@ namespace ClassLibrary
     {
         //add logger instance
         private static readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
-        public static void CopyFiles(string targetFile, string destination)
+        public static void CopyFiles(string targetFile, string destination, string _readonly="1")
         {
             try
             {
                 logger.Info($"copying file {targetFile} to {destination}");
                 //copy target file to destination, add filename to directory because of file.copy function
-                File.Copy(targetFile, destination + @"\" + Path.GetFileName(targetFile));
+                if (_readonly == "1")
+                {
+                    File.Copy(targetFile, destination + @"\" + Path.GetFileName(targetFile));
+                }
+                
                 //create "copied" file per mkv file make sure it doesnt get copied again.
                 using (FileStream fs = File.Create(targetFile + @"copied"))
                 {
